use snapshot writer for everything
