+++ query.go
@@ -6,64 +6,8 @@ import (
 	"fmt"
 	"github.com/urfave/cli"
 	"google.golang.org/api/iterator"
-	"io"
-	"log"
 	"strings"
 )
-
-// wrapper to stream the json serialized results
-type displayItemWriter struct {
-	isFirst bool
-	writer  *io.Writer
-}
-
-func newDisplayItemWriter(writer *io.Writer) displayItemWriter {
-	return displayItemWriter{true, writer}
-}
-
-func (d *displayItemWriter) Write(doc *firestore.DocumentSnapshot, extendedJson bool) error {
-	if d.isFirst {
-		_, err := fmt.Fprintln(*d.writer, "[")
-		if err != nil {
-			return err
-		}
-		d.isFirst = false
-	} else {
-		_, err := fmt.Fprintln(*d.writer, ",")
-		if err != nil {
-			return err
-		}
-	}
-
-	var displayItem = make(map[string]interface{})
-
-	displayItem["ID"] = doc.Ref.ID
-	displayItem["CreateTime"] = doc.CreateTime
-	displayItem["ReadTime"] = doc.ReadTime
-	displayItem["UpdateTime"] = doc.UpdateTime
-	displayItem["Data"] = doc.Data()
-
-	jsonString, err := marshallData(displayItem, extendedJson)
-
-	if err != nil {
-		return err
-	}
-
-	_, err = fmt.Fprintln(*d.writer, jsonString)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (d *displayItemWriter) Close() {
-	if !d.isFirst {
-		_, err := fmt.Fprintln(*d.writer, "]")
-		if err != nil {
-			log.Panicf("Could not write finishing part of results. %v", err)
-		}
-	}
-}
 
 func getDir(name string) firestore.Direction {
 	if name == "DESC" {
